more exercises

+++ exercices/019/solution.py
@@ -8,5 +8,5 @@ import sys
 if len(sys.argv) < 3:
     print("usage: python3 solution.py OP1 OP2")
 else:
-    var = float(sys.argv[1]) + float(sys.argv[2])
+    var = int(sys.argv[1]) + int(sys.argv[2])
     print(var)
